+++ extensions/caschys-blog/src/index.tsx
@@ -1,3 +1,1 @@
-// src/index.tsx
-
 export { default } from "./latest-articles";